backend/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.30.6
websockets==13.1
httpx==0.27.2
chromadb==1.0.0
pyyaml==6.0.2
python-dotenv==1.0.1
